requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmap==0.6.2
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.7.0
gdown==5.2.0
grpcio==1.73.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imgviz==1.8.0
Jinja2==3.1.6
kiwisolver==1.4.9
labelme==5.10.1
lazy_loader==0.4
loguru==0.7.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
opencv-python==4.12.0.88
osam==0.2.5
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.12.20
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.6.2
Werkzeug==3.1.3
